+++ src/components/Search/CancelAndCompareButtons.tsx
@@ -1,0 +1,47 @@
+import Button from '@mui/material/Button';
+import { style } from 'typestyle';
+
+import { Strings } from '../../resources/Strings';
+import { Spacing } from '../../styles/Spacing';
+
+interface CompareButtonProps {
+  label: string;
+  hasCancelButton: boolean;
+  onCancel: () => void;
+}
+
+export default function CompareButton({
+  label,
+  hasCancelButton,
+  onCancel,
+}: CompareButtonProps) {
+  const cancelText = Strings.components.searchDefault.sharedCollasped.cancel;
+
+  const cancelCompareStyles = style({
+    display: 'flex',
+    alignItems: 'center',
+    $nest: {
+      '.cancel-button': {
+        marginRight: `${Spacing.Small}px`,
+      },
+    },
+  });
+
+  return (
+    <div className={` ${cancelCompareStyles} cancel-compare`}>
+      {hasCancelButton && (
+        <Button
+          className={`cancel-button btn-all`}
+          name='cancel-button'
+          color='secondary'
+          onClick={onCancel}
+        >
+          {cancelText}
+        </Button>
+      )}
+      <Button id='compare-button' color='primary' type='submit'>
+        {label}
+      </Button>
+    </div>
+  );
+}
